Preserve view state for untitled files between restarts (fixes #16588)
