+++ src/vs/workbench/parts/files/browser/editors/textFileEditor.ts
@@ -10,9 +10,8 @@ import errors = require('vs/base/common/errors');
 import { toErrorMessage } from 'vs/base/common/errorMessage';
 import types = require('vs/base/common/types');
 import paths = require('vs/base/common/paths');
-import { IEditorViewState, IEditorOptions } from 'vs/editor/common/editorCommon';
+import { IEditorOptions } from 'vs/editor/common/editorCommon';
 import { Action } from 'vs/base/common/actions';
-import { Scope } from 'vs/workbench/common/memento';
 import { VIEWLET_ID, TEXT_FILE_EDITOR_ID } from 'vs/workbench/parts/files/common/files';
 import { ITextFileEditorModel, ITextFileService } from 'vs/workbench/services/textfile/common/textfiles';
 import { BaseTextEditor } from 'vs/workbench/browser/parts/editor/textEditor';
@@ -32,14 +31,6 @@ import { IInstantiationService } from 'vs/platform/instantiation/common/instanti
 import { IMessageService, CancelAction } from 'vs/platform/message/common/message';
 import { IWorkbenchEditorService } from 'vs/workbench/services/editor/common/editorService';
 import { IThemeService } from 'vs/workbench/services/themes/common/themeService';
-
-const TEXT_EDITOR_VIEW_STATE_PREFERENCE_KEY = 'textEditorViewState';
-
-interface ITextEditorViewState {
-	0?: IEditorViewState;
-	1?: IEditorViewState;
-	2?: IEditorViewState;
-}
 
 /**
  * An implementation of editor for file system resources.
@@ -72,7 +63,7 @@ export class TextFileEditor extends BaseTextEditor {
 	private onFilesChanged(e: FileChangesEvent): void {
 		const deleted = e.getDeleted();
 		if (deleted && deleted.length) {
-			this.clearTextEditorViewState(this.storageService, deleted.map((d) => d.resource.toString()));
+			this.clearTextEditorViewState(deleted.map(d => d.resource.toString()));
 		}
 	}
 
@@ -114,7 +105,7 @@ export class TextFileEditor extends BaseTextEditor {
 
 		// Remember view settings if input changes
 		if (oldInput) {
-			this.saveTextEditorViewState(this.storageService, oldInput.getResource().toString());
+			this.saveTextEditorViewState(oldInput.getResource().toString());
 		}
 
 		// Different Input (Reload)
@@ -146,7 +137,7 @@ export class TextFileEditor extends BaseTextEditor {
 			textEditor.setModel(textFileModel.textEditorModel);
 
 			// Always restore View State if any associated
-			const editorViewState = this.loadTextEditorViewState(this.storageService, this.getInput().getResource().toString());
+			const editorViewState = this.loadTextEditorViewState(this.getInput().getResource().toString());
 			if (editorViewState) {
 				textEditor.restoreViewState(editorViewState);
 			}
@@ -155,7 +146,7 @@ export class TextFileEditor extends BaseTextEditor {
 			if (options && types.isFunction((<TextEditorOptions>options).apply)) {
 				(<TextEditorOptions>options).apply(textEditor);
 			}
-		}, (error) => {
+		}, error => {
 
 			// In case we tried to open a file inside the text editor and the response
 			// indicates that this is not a text file, reopen the file through the binary
@@ -226,62 +217,11 @@ export class TextFileEditor extends BaseTextEditor {
 		return options;
 	}
 
-	/**
-	 * Saves the text editor view state under the given key.
-	 */
-	private saveTextEditorViewState(storageService: IStorageService, key: string): void {
-		const memento = this.getMemento(storageService, Scope.WORKSPACE);
-		let textEditorViewStateMemento = memento[TEXT_EDITOR_VIEW_STATE_PREFERENCE_KEY];
-		if (!textEditorViewStateMemento) {
-			textEditorViewStateMemento = Object.create(null);
-			memento[TEXT_EDITOR_VIEW_STATE_PREFERENCE_KEY] = textEditorViewStateMemento;
-		}
-
-		const editorViewState = this.getControl().saveViewState();
-
-		let fileViewState: ITextEditorViewState = textEditorViewStateMemento[key];
-		if (!fileViewState) {
-			fileViewState = Object.create(null);
-			textEditorViewStateMemento[key] = fileViewState;
-		}
-
-		if (typeof this.position === 'number') {
-			fileViewState[this.position] = editorViewState;
-		}
-	}
-
-	/**
-	 * Clears the text editor view state under the given key.
-	 */
-	private clearTextEditorViewState(storageService: IStorageService, keys: string[]): void {
-		const memento = this.getMemento(storageService, Scope.WORKSPACE);
-		const textEditorViewStateMemento = memento[TEXT_EDITOR_VIEW_STATE_PREFERENCE_KEY];
-		if (textEditorViewStateMemento) {
-			keys.forEach(key => delete textEditorViewStateMemento[key]);
-		}
-	}
-
-	/**
-	 * Loads the text editor view state for the given key and returns it.
-	 */
-	private loadTextEditorViewState(storageService: IStorageService, key: string): IEditorViewState {
-		const memento = this.getMemento(storageService, Scope.WORKSPACE);
-		const textEditorViewStateMemento = memento[TEXT_EDITOR_VIEW_STATE_PREFERENCE_KEY];
-		if (textEditorViewStateMemento) {
-			const fileViewState: ITextEditorViewState = textEditorViewStateMemento[key];
-			if (fileViewState) {
-				return fileViewState[this.position];
-			}
-		}
-
-		return null;
-	}
-
 	public clearInput(): void {
 
 		// Keep editor view state in settings to restore when coming back
 		if (this.input) {
-			this.saveTextEditorViewState(this.storageService, this.getInput().getResource().toString());
+			this.saveTextEditorViewState(this.getInput().getResource().toString());
 		}
 
 		// Clear Model
@@ -295,7 +235,7 @@ export class TextFileEditor extends BaseTextEditor {
 
 		// Save View State
 		if (this.input) {
-			this.saveTextEditorViewState(this.storageService, this.getInput().getResource().toString());
+			this.saveTextEditorViewState(this.getInput().getResource().toString());
 		}
 
 		// Call Super
